qim-prod-sdk-pkg: add mltflite plugin to package list
Issue: gstreamer mltflite plugin was not getting packed on some
machines due to timing issue. When mltflite plugin is compiled before
qim-sdk, it will be picked by qim-sdk for packaging as we search with
gstreamer string during packaging in qim-sdk. When this is not the case
mltflite ipk will not be part of the qim-sdk tar and hence won't be
part of qim-prod-sdk as well.

Fix: add mltflite plugins to the package list to be package via
qim-prod-sdk-pkg as we enable mltflite plugin and other ml plugins
from this layer.

Change-Id: I49ca5cfd8cadefa18ea58c747a4efcc2a54c1152